+++ packages/types/src/elements/input/index.ts
@@ -5,9 +5,12 @@ import {
 } from "../..";
 import { StringRule } from "../../rules";
 import { GeneralStyleMetadata } from "../../styles";
+import { IconType } from "../icon";
+import { ImageType } from "../image";
+import { TextType } from "../text";
 
 export type InputType = GenericLezzformElement<
-  GenericFieldAttributes,
+  GenericFieldAttributes & InputAttributesType,
   InputStyleType
 > & {
   type: "Input";
@@ -15,6 +18,34 @@ export type InputType = GenericLezzformElement<
   rule: InputRuleType;
 };
 
-export type InputStyleType = GeneralStyleMetadata;
+export type InputIconType = Pick<IconType, "type" | "attributes">;
+export type InputImageType = Pick<ImageType, "type" | "attributes">;
+
+export type InputAdornmentType = {
+  icon: InputIconType | InputImageType;
+  text: Omit<TextType, "id">;
+};
+
+export type InputAttributesType = {
+  adornment?: {
+    type?: "prefix" | "suffix";
+    prefix?: Partial<InputAdornmentType>;
+    suffix?: Partial<InputAdornmentType>;
+  };
+};
+
+export type InputStyleType = GeneralStyleMetadata<InputStyleMetadataValuesType>;
+
+export type InputStyleMetadataValuesType = {
+  input: Record<string, unknown>;
+  prefixAdornment: {
+    icon?: Record<string, unknown>;
+    text?: Record<string, unknown>;
+  };
+  suffixAdornment: {
+    icon?: Record<string, unknown>;
+    text?: Record<string, unknown>;
+  };
+};
 
 export type InputRuleType = StringRule;